Simplify worker stages

+++ followthegrant/worker/tasks.py
@@ -5,7 +5,6 @@ from functools import cache
 from followthemoney.util import make_entity_id
 from structlog import get_logger
 
-from followthegrant import settings
 from followthegrant.dedupe import explode_triples
 from followthegrant.parse import parse
 from followthegrant.store import get_dataset, get_store
@@ -35,9 +34,7 @@ def get_stage(dataset: str, job_id: str, queue: str) -> Stage:
 
 
 PARSE = "parse"
-DELETE_SOURCE = "delete-source"
 WRITE_FTM = "write-ftm"
-AUTHOR_TRIPLES = "author-triples"
 WRITE_AUTHOR_TRIPLES = "write-author-triples"
 
 
@@ -49,16 +46,17 @@ def op_parse(payload):
     # rabbitmq message size limit
     for proxy in parse(fpath, parser, dataset):
         yield WRITE_FTM, {**payload, **{"data": [proxy.to_dict()]}}
-        yield AUTHOR_TRIPLES, {**payload, **{"data": [proxy.to_dict()]}}
-    # if configured, delete source file to decrease disk space
-    yield DELETE_SOURCE, payload
-
-
-def op_delete_source(payload):
-    try:
-        os.remove(get_path(payload["fpath"]))
-    except (OSError, FileNotFoundError):
-        pass
+        triples = set()
+        for triple in explode_triples(proxy):
+            triples.add(triple)
+        if triples:
+            yield WRITE_AUTHOR_TRIPLES, {**payload, **{"data": list(triples)}}
+    # if configured, delete source file to free up disk space
+    if payload.get("delete_source"):
+        try:
+            os.remove(get_path(payload["fpath"]))
+        except (OSError, FileNotFoundError):
+            pass
 
 
 def op_write_ftm(dataset, entities):
@@ -69,14 +67,6 @@ def op_write_ftm(dataset, entities):
     bulk.flush()
 
 
-def op_author_triples(payload):
-    triples = set()
-    for proxy in payload["data"]:
-        for triple in explode_triples(proxy):
-            triples.add(triple)
-    yield WRITE_AUTHOR_TRIPLES, {**payload, **{"data": list(triples)}}
-
-
 def op_write_author_triples(dataset, rows):
     rows = (r + [dataset] for r in rows)
     store = get_store()
@@ -85,16 +75,14 @@ def op_write_author_triples(dataset, rows):
 
 QUEUES = {
     # stage: (func, batch_size, *next_stages)
-    PARSE: (op_parse, 1, WRITE_FTM, AUTHOR_TRIPLES, DELETE_SOURCE),
-    DELETE_SOURCE: (op_delete_source, 10_000),
+    PARSE: (op_parse, 1, WRITE_FTM, WRITE_AUTHOR_TRIPLES),
     WRITE_FTM: (op_write_ftm, 10_000),
-    AUTHOR_TRIPLES: (op_author_triples, 10_000, WRITE_AUTHOR_TRIPLES),
     WRITE_AUTHOR_TRIPLES: (op_write_author_triples, 10_000),
 }
 
 
-class TaskAggregator:
-    """handle a collection of identical tasks"""
+class BulkWriter:
+    """collect tasks to bulk write"""
 
     def __init__(self, consumer, stage: Stage):
         self.is_flushing = False
@@ -104,8 +92,6 @@ class TaskAggregator:
         func, batch_size, *next_queues = QUEUES[stage.queue]
         self.func = func
         self.batch_size = batch_size
-        self.next_queues = next_queues
-        self.is_writer = self.stage.queue.startswith("write")
         self.last_activity = time.time()
 
     def add(self, tag, payload):
@@ -122,39 +108,12 @@ class TaskAggregator:
             to_write = []
 
             for delivery_tag, payload in self.tasks:
-                next_queues = set(payload.get("allowed_queues", QUEUES.keys())) & set(
-                    self.next_queues
-                )
-                try:
-                    acked = False
-                    if self.is_writer:
-                        for item in payload["data"]:
-                            to_write.append(item)
-                    else:
-                        res = self.func(payload)
-                        if res is not None:
-                            for queue, payload in res:
-                                if not acked:
-                                    # ack message already, as some tasks take very long
-                                    self.consumer.ack(delivery_tag)
-                                    done += 1
-                                    acked = True
-                                if queue in next_queues:
-                                    self.consumer.dispatch(queue, payload)
+                for item in payload["data"]:
+                    to_write.append(item)
+                self.consumer.ack(delivery_tag)
+                done += 1
 
-                except Exception as e:
-                    msg = f'{self.stage} : {e} : {payload["fpath"]}'
-                    log.error(msg)
-                    if settings.DEBUG:
-                        log.exception(msg, payload=payload, exception=e)
-                    errors += 1
-                    self.consumer.nack(delivery_tag, requeue=False)
-
-                if not acked:
-                    self.consumer.ack(delivery_tag)
-                    done += 1
-
-            if self.is_writer and len(to_write):
+            if len(to_write):
                 try:
                     self.func(self.stage.dataset, to_write)
                 except Exception as e:
